+++ README.md
@@ -24,6 +24,7 @@
 
 - [How to install](#how-to-install)
 - [Command Line Usage](#command-line-usage)
+- [The bananafile.json](#the-bananafile)
 - [Module Usage](#module-usage)
 - [Features](#features)
 - [Example](#example)
@@ -90,6 +91,32 @@ $ banana --help
 
 <hr>
 
+## The bananafile
+
+Create a file called `bananafile.json` in the directory where you will run the command `$ banana` and configure as need.
+
+```json
+{
+	"bnnSize" : true,
+	"bnnPosition" : true,
+	"bnnGradient" : true,
+	"bnnVariable" : true,
+	"bnnImport" : true,
+	"bnnAlign" : true,
+	"bnnWidth" : true,
+	"bnnHeight" : true,
+	"bnnCol" : true,
+	"bnnRow" : true,
+	"bnnBox" : true,
+	"bnnFunction" : true,
+	"compress" : false
+}
+```
+
+*All features have `true` as default value, except the `compress`.*
+
+<hr>
+
 ## Module Usage
 
 ```js
@@ -97,19 +124,19 @@ const inputBananaCode = '.a {bnn-size: 50px;}';
 
 // Features injection
 const config = {};
-config.bnnSize = true; // Default: false
-config.bnnPosition = true; // Default: false
-config.bnnGradient = true; // Default: false
-config.bnnVariable = true; // Default: false
-config.bnnImport = true; // Default: false
-config.bnnAlign = true; // Default: false
-config.bnnWidth = true; // Default: false
-config.bnnHeight = true; // Default: false
-config.bnnCol = true; // Default: false
-config.bnnRow = true; // Default: false
-config.bnnBox = true; // Default: false
-config.bnnFunction = true; // Default: false
-config.compress = true; // Default: false
+config.bnnSize = true;
+config.bnnPosition = true;
+config.bnnGradient = true;
+config.bnnVariable = true;
+config.bnnImport = true;
+config.bnnAlign = true;
+config.bnnWidth = true;
+config.bnnHeight = true;
+config.bnnCol = true;
+config.bnnRow = true;
+config.bnnBox = true;
+config.bnnFunction = true;
+config.compress = true;
 
 const Banana = require('banana')(config);
 
@@ -134,6 +161,7 @@ console.log(output); // .a {width: 50px; height: 50px;}
 - [Module Bundler](docs/module-bundler.md) with native `@import` syntax.
 - Global [variables](docs/variables.md) with native custom properties syntax.
 - Create reusable functions with [@function](docs/functions.md).
+- Configure your build with [bananafile.json](the-bananafile).
 - Minify/Compress the generated CSS.
 
 View all [features docs here](docs/index.md).
