+++ bin/defaultConfig.js
@@ -1,0 +1,17 @@
+const config = {};
+
+config.bnnSize = true;
+config.bnnPosition = true;
+config.bnnGradient = true;
+config.bnnVariable = true;
+config.bnnImport = true;
+config.bnnAlign = true;
+config.bnnWidth = true;
+config.bnnHeight = true;
+config.bnnCol = true;
+config.bnnRow = true;
+config.bnnBox = true;
+config.bnnFunction = true;
+config.compress = false;
+
+module.exports = config;
