+++ src/banana.js
@@ -28,7 +28,7 @@ const Banana = (config) => {
 
       // Search for all the @import and generate a single AST
       ast.findAllImport((urlForImport, index) => {
-        if (config.bnnImport) {
+        if (config.bnnImport || config.bnnImport === undefined) {
           const bnnImport = require('../src/core/bnnImport.js');
           bnnImport(urlForImport, ast, index, inputPath);
         }
@@ -36,7 +36,7 @@ const Banana = (config) => {
 
       // Search for all global variables and compile
       ast.findAllRulesBySelectors(':root', (rule, index) => {
-        if (config.bnnVariable) {
+        if (config.bnnVariable || config.bnnVariable === undefined) {
           require('../src/core/bnnVariable.js')(ast, rule, index);
         }
       });
@@ -44,39 +44,39 @@ const Banana = (config) => {
       // Search for all banana properties and compile
       ast.findAllRulesByType('rule', (rule) => {
 
-        if (config.bnnSize) {
+        if (config.bnnSize || config.bnnSize === undefined) {
           require('../src/core/bnnSize.js')(rule);
         }
 
-        if (config.bnnPosition) {
+        if (config.bnnPosition || config.bnnPosition === undefined) {
           require('../src/core/bnnPosition.js')(rule);
         }
 
-        if (config.bnnGradient) {
+        if (config.bnnGradient || config.bnnGradient === undefined) {
           require('../src/core/bnnGradient.js')(rule);
         }
 
-        if (config.bnnAlign) {
+        if (config.bnnAlign || config.bnnAlign === undefined) {
           require('../src/core/bnnAlign.js')(rule);
         }
 
-        if (config.bnnWidth) {
+        if (config.bnnWidth || config.bnnWidth === undefined) {
           require('../src/core/bnnWidth.js')(rule);
         }
 
-        if (config.bnnHeight) {
+        if (config.bnnHeight || config.bnnHeight === undefined) {
           require('../src/core/bnnHeight.js')(rule);
         }
 
-        if (config.bnnCol) {
+        if (config.bnnCol || config.bnnCol === undefined) {
           require('../src/core/bnnCol.js')(rule);
         }
 
-        if (config.bnnRow) {
+        if (config.bnnRow || config.bnnRow === undefined) {
           require('../src/core/bnnRow.js')(rule);
         }
 
-        if (config.bnnBox) {
+        if (config.bnnBox || config.bnnBox === undefined) {
           require('../src/core/bnnBox.js')(rule);
         }
 
@@ -86,7 +86,7 @@ const Banana = (config) => {
       ast.moonWalkAllRules((rule, index) => {
         const isFunction = /\@function\ /.test(rule.selectors);
         if (isFunction) {
-          if (config.bnnFunction) {
+          if (config.bnnFunction || config.bnnFunction === undefined) {
             require('../src/core/bnnFunction.js')(rule, ast, index);
           }
         }
